Renamed utils.py to bcolors.py
